+++ NEWS.md
@@ -1,3 +1,58 @@
+## meta, version 4.19-0 (2021-mm-dd)
+
+### Major changes
+
+* Subgroup analysis for three-level model fully implemented
+
+* New default for forest plots to show results of test for subgroup
+  differences in meta-analyses with subgroups
+
+* Calculation of weights for three-level random effects model using
+  weights.rma.mv() with argument type = "rowsum" from R package
+  **metafor**
+
+* Print study label provided by argument 'studlab' for meta-analysis
+  with a single study
+
+* Total number of observations and events printed in summaries (if
+  available)
+
+### Bug fixes
+
+* metagen():
+  - treatment estimates for three-level models with subgroups were not
+    based on common between-study variance despite argument tau.common
+    = TRUE
+
+* metareg():
+  - use rma.mv() from R package **metafor** for three-level models
+    instead of rma.uni()
+
+### User-visible changes
+
+* metabin(), metacont(), metacor(), metacr(), metagen(), metagen(),
+  metainc(), metamean(), metaprop(), metarate():
+  - new argument 'test.subgroup' to print results of test for subgroup
+    differences
+
+* print.meta():
+  - for three-level models, column with grouping information added to
+    study details
+
+* metagen():
+  - default for estimation of between-study variance has changed for
+    three-level models with subgroups, i.e., tau2 is allowed to be
+    different in subgroups by default
+
+### Internal changes
+
+* metagen():
+  - new variable '.idx' with running index in meta-analysis data set
+    (list element 'data')
+  - new logical list element 'three.level' indicating whether
+    three-level model was used
+
+
 ## meta, version 4.18-2 (2021-06-11)
 
 ### Bug fixes
@@ -1019,7 +1074,7 @@
 * New function metabind() to combine meta-analysis objects, e.g. to
   generate a forest plot with results of several subgroup analyses
 
-* Subgroup analyses implemented for generalised linear mixed models
+* Subgroup analysis implemented for generalised linear mixed models
   (GLMMs) with and without assumption of common between-study variance
   (arguments 'byvar' and 'tau.common')